// src/app/core/restaurant_service/model/timeRange.ts
/**
 * OpenAPI definition
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface TimeRange { 
    timeRangeCode?: string;
    openingHour?: number;
    openingMinutes?: number;
    closingHour?: number;
    closingMinutes?: number;
}